CORE-1140 Refactor Notifications TableView into separate file.
